refactor: Formas y métodos de pago ahora manejan la lógica de carga de datos desde composables
